Move common code to root and end step under src
This is the beginning diff. Moved the API server & package.json to the root and created new `src/` folder which will contain all of the steps (including `workshop`). This commit just moves `end` to `src/end`
