+++ src/end/components/EmailView.js
